moves packages analyzers and groupers into steps
