+++ src/dimcat/__init__.py
@@ -14,10 +14,10 @@ finally:
 
 from .base import Data, DimcatConfig, DimcatObject, DimcatSchema, get_class, get_schema
 from .dataset.base import Dataset, DimcatPackage
-from .groupers.base import CustomPieceGrouper
 from .resources.base import DimcatIndex, DimcatResource, PieceIndex
 from .resources.features import FeatureName
 from .steps.base import PipelineStep
+from .steps.groupers import CustomPieceGrouper
 from .steps.pipelines import Pipeline
 
 logger = logging.getLogger(__name__)
