+++ src/dimcat/steps/__init__.py
@@ -1,7 +1,9 @@
 import logging
 
+from .analyzers import Analyzer, Counter
 from .base import PipelineStep
 from .extractors import FeatureExtractor
+from .groupers import CustomPieceGrouper, Grouper
 from .pipelines import Pipeline
 
 logger = logging.getLogger(__name__)
