+++ src/dimcat/steps/analyzers/__init__.py
@@ -1,5 +1,6 @@
 import logging
 
+from .base import Analyzer
 from .counters import Counter
 
 logger = logging.getLogger(__name__)
