+++ src/dimcat/steps/groupers/__init__.py
@@ -1,5 +1,5 @@
 import logging
 
-from .base import CustomPieceGrouper
+from .base import CustomPieceGrouper, Grouper
 
 logger = logging.getLogger(__name__)
