+++ tests/test_base.py
@@ -10,8 +10,6 @@ from typing import List, Tuple, Type
 import frictionless as fl
 import pandas as pd
 import pytest
-from dimcat.analyzers import Counter
-from dimcat.analyzers.base import Analyzer
 from dimcat.base import (
     Data,
     DimcatConfig,
@@ -24,6 +22,7 @@ from dimcat.base import (
 from dimcat.dataset.base import DimcatCatalog, DimcatPackage
 from dimcat.resources.base import DimcatResource
 from dimcat.resources.features import Notes
+from dimcat.steps.analyzers import Analyzer, Counter
 from dimcat.steps.base import PipelineStep
 from marshmallow import ValidationError, fields
 from marshmallow.class_registry import _registry as MM_REGISTRY
